+++ src/lib/world/worlddc.h
@@ -568,11 +568,11 @@ namespace madness {
             ////(acc_gpu->second->aggregateParams(cb));
                //Need to change this to not overwrite the object, but add to its vector 
 
-            ConcurrentHashMap<long, ComputeBase *>::iterator gpu_it;
-            ConcurrentHashMap<long, ComputeBase *>::iterator gpu_end = this->world.gpu_hash.end();
+            ConcurrentHashMap<unsigned long long, ComputeBase *>::iterator gpu_it;
+            ConcurrentHashMap<unsigned long long, ComputeBase *>::iterator gpu_end = this->world.gpu_hash.end();
             
             this->world.gpu_hashlock.lock();
-            gpu_it = this->world.gpu_hash.find((long)((void *)memfun2));
+            gpu_it = this->world.gpu_hash.find((unsigned long long)((void *)memfun2));
             //gpu_it = this->world.gpu_hash.find(0);
             if (gpu_it != gpu_end){ 
                 (*gpu_it).second->add(cd->inObj.at(0));
@@ -581,7 +581,7 @@ namespace madness {
 
             }
             else{
-                this->world.gpu_hash.insert(std::pair<long, ComputeBase *>((long)((void *)memfun2), cb));
+                this->world.gpu_hash.insert(std::pair<unsigned long long, ComputeBase *>((unsigned long long)((void *)memfun2), cb));
                 //this->world.gpu_hash.insert(std::pair<long, ComputeBase *>(0, cb));
             }
 
@@ -633,11 +633,11 @@ namespace madness {
             ////(acc_gpu->second->aggregateParams(cb));
                //Need to change this to not overwrite the object, but add to its vector 
 
-            ConcurrentHashMap<long, ComputeBase *>::iterator gpu_it;
-            ConcurrentHashMap<long, ComputeBase *>::iterator gpu_end = this->world.gpu_hash.end();
+            ConcurrentHashMap<unsigned long long, ComputeBase *>::iterator gpu_it;
+            ConcurrentHashMap<unsigned long long, ComputeBase *>::iterator gpu_end = this->world.gpu_hash.end();
             
             this->world.gpu_hashlock.lock();
-            gpu_it = this->world.gpu_hash.find((long)(/*(void *)*/reinterpret_cast<void *>(memfun2)));
+            gpu_it = this->world.gpu_hash.find((unsigned long long)(/*(void *)*/reinterpret_cast<void *>(memfun2)));
             //gpu_it = this->world.gpu_hash.find(0);
             if (gpu_it != gpu_end){ 
                 (*gpu_it).second->add(cd->inObj.at(0));
@@ -646,7 +646,7 @@ namespace madness {
 
             }
             else{
-                this->world.gpu_hash.insert(std::pair<long, ComputeBase *>((long)(/*(void *)*/reinterpret_cast<void *>(memfun2)), cb));
+                this->world.gpu_hash.insert(std::pair<unsigned long long, ComputeBase *>((unsigned long long)(/*(void *)*/reinterpret_cast<void *>(memfun2)), cb));
                 //this->world.gpu_hash.insert(std::pair<long, ComputeBase *>(0, cb));
             }
 
@@ -697,8 +697,8 @@ namespace madness {
             ////(acc_gpu->second->aggregateParams(cb));
                //Need to change this to not overwrite the object, but add to its vector 
 
-            ConcurrentHashMap<long, ComputeBase *>::iterator gpu_it;
-            ConcurrentHashMap<long, ComputeBase *>::iterator gpu_end = this->world.gpu_hash.end();
+            ConcurrentHashMap<unsigned long long, ComputeBase *>::iterator gpu_it;
+            ConcurrentHashMap<unsigned long long, ComputeBase *>::iterator gpu_end = this->world.gpu_hash.end();
             
             this->world.gpu_hashlock.lock();
             //gpu_it = this->world.gpu_hash.find((long)(&memfun2));
@@ -711,7 +711,7 @@ namespace madness {
             }
             else{
                 //this->world.gpu_hash.insert(std::pair<long, ComputeBase *>((long)(&memfun2), cb));
-                this->world.gpu_hash.insert(std::pair<long, ComputeBase *>(0, cb));
+                this->world.gpu_hash.insert(std::pair<unsigned long long, ComputeBase *>(0, cb));
             }
 
             this->world.taskq.incNRegistered();
